--- skeleton/client/index.js
@@ -1,5 +1,0 @@
-import { Assembly } from "../client/components/Assembly.js"
-import KeydownEvenet from "../client/module/KeydownEvent.js"
-
-
-KeydownEvenet("Enter", "button")
