+++ skeleton/client/public/js/components/LiComponent.js
