+++ skeleton/client/public/js/components/TagComponent.js
