+++ skeleton/client/public/js/index.js
@@ -1,0 +1,5 @@
+import { Assembly } from "./components/Assembly.js"
+import KeydownEvenet from "./module/KeydownEvent.js"
+
+
+KeydownEvenet("Enter", "button")
